Infrastruktura/functions/fa-worker/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-identity==1.17.1
azure-servicebus==7.14.2
azure-functions>=1.20.0
psycopg[binary]>=3.1,<3.2